Introducing concept of backup flags for Griddata
Griddata previously defaulted to an "averaging" scheme when
an interpolation method failed. This could lead to the user
believing it worked when it really is just an average value.

By default, griddata will now return a flag value if a method
does not work. The user can specify a backup flag set as desiered